% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{cat_estimation.default}
\alias{cat_estimation.default}
\title{cat_estimation_default}
\usage{
\method{cat_estimation}{default}(estimation, call, level = 0.95)
}
\arguments{
\item{estimation}{Estimation}

\item{call}{Call}
}
\description{
cat_estimation_default
}
\keyword{internal}
