% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecipeCategory.R
\name{default_categories}
\alias{default_categories}
\title{Default recipe categories}
\usage{
default_categories()
}
\value{
List of RecipeCategory objects
}
\description{
Returns a list of standard built-in
categories for recipe classification.
}
\examples{
cats <- default_categories()
vapply(cats, function(c) c$name, character(1))
}
\seealso{
Other tidy-api:
\code{\link{RecipeCategory-from_list}},
\code{\link{RecipeCertification}},
\code{\link{RecipeUser}},
\code{\link[=add_category]{add_category()}},
\code{\link[=certify_recipe]{certify_recipe()}},
\code{\link[=filter_recipes]{filter_recipes()}},
\code{\link[=filter_workflows]{filter_workflows()}},
\code{\link[=find_workflows_for_recipe]{find_workflows_for_recipe()}},
\code{\link[=list_recipes]{list_recipes()}},
\code{\link[=list_workflows]{list_workflows()}},
\code{\link[=rank_recipes]{rank_recipes()}},
\code{\link[=rank_workflows]{rank_workflows()}},
\code{\link[=recipe_category]{recipe_category()}},
\code{\link[=recipe_certification]{recipe_certification()}},
\code{\link[=recipe_user]{recipe_user()}},
\code{\link[=remove_category]{remove_category()}},
\code{\link[=search_recipes]{search_recipes()}},
\code{\link[=search_workflows]{search_workflows()}},
\code{\link[=set_user_info]{set_user_info()}},
\code{\link[=set_version]{set_version()}}
}
\concept{tidy-api}
