% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecipeAPI.R
\name{get_backend}
\alias{get_backend}
\title{Get recipe backend}
\usage{
get_backend()
}
\value{
RecipeBackend object
}
\description{
Returns the currently configured recipe backend.
Defaults to "api" if not configured.
}
\examples{
set_backend("local", path = tempfile(fileext = ".json"))
backend <- get_backend()
backend
}
\seealso{
Other backends:
\code{\link[=get_workflow_backend]{get_workflow_backend()}},
\code{\link[=set_backend]{set_backend()}},
\code{\link[=set_workflow_backend]{set_workflow_backend()}}
}
\concept{backends}
