% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_engine.R
\name{get_engine}
\alias{get_engine}
\title{Get the current survey data engine}
\usage{
get_engine()
}
\value{
Character vector with the name of the configured engine.
}
\description{
Retrieves the currently configured engine
for loading surveys from system options or environment
variables.
}
\examples{
get_engine()
}
\seealso{
Other options:
\code{\link[=lazy_default]{lazy_default()}},
\code{\link[=set_engine]{set_engine()}},
\code{\link[=set_lazy_processing]{set_lazy_processing()}},
\code{\link[=set_use_copy]{set_use_copy()}},
\code{\link[=show_engines]{show_engines()}},
\code{\link[=use_copy_default]{use_copy_default()}}
}
\concept{options}
\keyword{engine}
