% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorkflowBackend.R
\name{get_workflow_backend}
\alias{get_workflow_backend}
\title{Get workflow backend}
\usage{
get_workflow_backend()
}
\value{
WorkflowBackend object
}
\description{
Returns the currently configured workflow backend.
Defaults to "local" if not configured.
}
\examples{
backend <- get_workflow_backend()
}
\seealso{
Other backends:
\code{\link[=get_backend]{get_backend()}},
\code{\link[=set_backend]{set_backend()}},
\code{\link[=set_workflow_backend]{set_workflow_backend()}}
}
\concept{backends}
