% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{has_design}
\alias{has_design}
\title{Check if survey has a design}
\usage{
has_design(svy)
}
\arguments{
\item{svy}{A Survey object.}
}
\value{
Logical.
}
\description{
Check if survey has a design
}
\examples{
svy <- survey_empty(type = "test", edition = "2023")
has_design(svy) # FALSE
}
\seealso{
Other survey-objects:
\code{\link{Survey}},
\code{\link[=cat_design]{cat_design()}},
\code{\link[=cat_design_type]{cat_design_type()}},
\code{\link[=get_data]{get_data()}},
\code{\link[=get_metadata]{get_metadata()}},
\code{\link[=has_recipes]{has_recipes()}},
\code{\link[=has_steps]{has_steps()}},
\code{\link[=is_baked]{is_baked()}},
\code{\link[=set_data]{set_data()}},
\code{\link[=survey_empty]{survey_empty()}},
\code{\link[=survey_to_data_frame]{survey_to_data_frame()}},
\code{\link[=survey_to_datatable]{survey_to_datatable()}},
\code{\link[=survey_to_tibble]{survey_to_tibble()}}
}
\concept{survey-objects}
