% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{is_baked}
\alias{is_baked}
\title{Check if all steps are baked}
\usage{
is_baked(svy)
}
\arguments{
\item{svy}{A Survey or RotativePanelSurvey object.}
}
\value{
Logical.
}
\description{
Returns TRUE when every step attached to the survey has been
executed (bake == TRUE), or when there are no steps.
}
\examples{
svy <- survey_empty(type = "test", edition = "2023")
is_baked(svy) # TRUE (no steps)
}
\seealso{
Other survey-objects:
\code{\link{Survey}},
\code{\link[=cat_design]{cat_design()}},
\code{\link[=cat_design_type]{cat_design_type()}},
\code{\link[=get_data]{get_data()}},
\code{\link[=get_metadata]{get_metadata()}},
\code{\link[=has_design]{has_design()}},
\code{\link[=has_recipes]{has_recipes()}},
\code{\link[=has_steps]{has_steps()}},
\code{\link[=set_data]{set_data()}},
\code{\link[=survey_empty]{survey_empty()}},
\code{\link[=survey_to_data_frame]{survey_to_data_frame()}},
\code{\link[=survey_to_datatable]{survey_to_datatable()}},
\code{\link[=survey_to_tibble]{survey_to_tibble()}}
}
\concept{survey-objects}
