% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_tidy_api.R
\name{list_workflows}
\alias{list_workflows}
\title{List all workflows}
\usage{
list_workflows()
}
\value{
List of all RecipeWorkflow objects.
}
\description{
List all workflows from the active workflow backend.
}
\examples{
set_workflow_backend("local", path = tempfile(fileext = ".json"))
all <- list_workflows()
length(all)

}
\seealso{
\code{\link{search_workflows}}, \code{\link{filter_workflows}}

Other tidy-api:
\code{\link{RecipeCategory}},
\code{\link{RecipeCertification}},
\code{\link{RecipeUser}},
\code{\link[=add_category]{add_category()}},
\code{\link[=certify_recipe]{certify_recipe()}},
\code{\link[=default_categories]{default_categories()}},
\code{\link[=filter_recipes]{filter_recipes()}},
\code{\link[=filter_workflows]{filter_workflows()}},
\code{\link[=find_workflows_for_recipe]{find_workflows_for_recipe()}},
\code{\link[=list_recipes]{list_recipes()}},
\code{\link[=rank_recipes]{rank_recipes()}},
\code{\link[=rank_workflows]{rank_workflows()}},
\code{\link[=recipe_category]{recipe_category()}},
\code{\link[=recipe_certification]{recipe_certification()}},
\code{\link[=recipe_user]{recipe_user()}},
\code{\link[=remove_category]{remove_category()}},
\code{\link[=search_recipes]{search_recipes()}},
\code{\link[=search_workflows]{search_workflows()}},
\code{\link[=set_user_info]{set_user_info()}},
\code{\link[=set_version]{set_version()}}
}
\concept{tidy-api}
