% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_lazy_processing}
\alias{set_lazy_processing}
\title{Set lazy processing}
\usage{
set_lazy_processing(lazy)
}
\arguments{
\item{lazy}{Logical. If TRUE, steps are deferred until
bake_steps() is called.}
}
\value{
Invisibly, the previous value.
}
\description{
Set lazy processing
}
\examples{
old <- lazy_default()
set_lazy_processing(FALSE)
lazy_default() # now FALSE
set_lazy_processing(old) # restore
}
\seealso{
Other options:
\code{\link[=get_engine]{get_engine()}},
\code{\link[=lazy_default]{lazy_default()}},
\code{\link[=set_engine]{set_engine()}},
\code{\link[=set_use_copy]{set_use_copy()}},
\code{\link[=show_engines]{show_engines()}},
\code{\link[=use_copy_default]{use_copy_default()}}
}
\concept{options}
\keyword{utils}
