% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe_tidy_api.R
\name{set_version}
\alias{set_version}
\title{Set version on a recipe}
\usage{
set_version(recipe, version)
}
\arguments{
\item{recipe}{A Recipe object.}

\item{version}{Character version string (e.g. \code{"2.0.0"}).}
}
\value{
The modified Recipe object.
}
\description{
Pipe-friendly function to set the version string on a Recipe.
}
\examples{
r <- recipe(
  name = "Example", user = "Test",
  svy = survey_empty(type = "ech", edition = "2023"),
  description = "Example recipe"
)
r <- r |> set_version("2.0.0")

}
\seealso{
Other tidy-api:
\code{\link{RecipeCategory}},
\code{\link{RecipeCertification}},
\code{\link{RecipeUser}},
\code{\link[=add_category]{add_category()}},
\code{\link[=certify_recipe]{certify_recipe()}},
\code{\link[=default_categories]{default_categories()}},
\code{\link[=filter_recipes]{filter_recipes()}},
\code{\link[=filter_workflows]{filter_workflows()}},
\code{\link[=find_workflows_for_recipe]{find_workflows_for_recipe()}},
\code{\link[=list_recipes]{list_recipes()}},
\code{\link[=list_workflows]{list_workflows()}},
\code{\link[=rank_recipes]{rank_recipes()}},
\code{\link[=rank_workflows]{rank_workflows()}},
\code{\link[=recipe_category]{recipe_category()}},
\code{\link[=recipe_certification]{recipe_certification()}},
\code{\link[=recipe_user]{recipe_user()}},
\code{\link[=remove_category]{remove_category()}},
\code{\link[=search_recipes]{search_recipes()}},
\code{\link[=search_workflows]{search_workflows()}},
\code{\link[=set_user_info]{set_user_info()}}
}
\concept{tidy-api}
