% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{survey_empty}
\alias{survey_empty}
\title{survey_empty}
\usage{
survey_empty(
  edition = NULL,
  type = NULL,
  weight = NULL,
  engine = NULL,
  psu = NULL,
  strata = NULL
)
}
\arguments{
\item{edition}{Edition of survey}

\item{type}{Type of survey}

\item{weight}{Weight of survey}

\item{engine}{Engine of survey}

\item{psu}{PSU variable or formula (optional)}

\item{strata}{Stratification variable name (optional)}
}
\value{
Survey object
}
\description{
Create an empty survey
}
\examples{
empty <- survey_empty(edition = "2023", type = "test")
empty_typed <- survey_empty(edition = "2023", type = "ech")
}
\seealso{
Other survey-objects:
\code{\link{Survey}},
\code{\link[=cat_design]{cat_design()}},
\code{\link[=cat_design_type]{cat_design_type()}},
\code{\link[=get_data]{get_data()}},
\code{\link[=get_metadata]{get_metadata()}},
\code{\link[=has_design]{has_design()}},
\code{\link[=has_recipes]{has_recipes()}},
\code{\link[=has_steps]{has_steps()}},
\code{\link[=is_baked]{is_baked()}},
\code{\link[=set_data]{set_data()}},
\code{\link[=survey_to_data_frame]{survey_to_data_frame()}},
\code{\link[=survey_to_datatable]{survey_to_datatable()}},
\code{\link[=survey_to_tibble]{survey_to_tibble()}}
}
\concept{survey-objects}
\keyword{survey}
