% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_time_pattern}
\alias{validate_time_pattern}
\title{Validate time pattern}
\usage{
validate_time_pattern(svy_type = NULL, svy_edition = NULL)
}
\arguments{
\item{svy_type}{Survey type (e.g. "ech").}

\item{svy_edition}{Survey edition string
(e.g. "2023", "2023-06").}
}
\value{
List with components: svy_type, svy_edition
(parsed), svy_periodicity.
}
\description{
Validate time pattern
}
\examples{
validate_time_pattern(svy_type = "ech", svy_edition = "2023")
validate_time_pattern(
  svy_type = "ech", svy_edition = "2023-06"
)
}
\seealso{
Other survey-loading:
\code{\link[=extract_time_pattern]{extract_time_pattern()}},
\code{\link[=group_dates]{group_dates()}},
\code{\link[=load_panel_survey]{load_panel_survey()}},
\code{\link[=load_survey]{load_survey()}},
\code{\link[=load_survey_example]{load_survey_example()}}
}
\concept{survey-loading}
\keyword{utils}
